backend/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
argcomplete==1.10.3
asgiref==3.8.1
audioread==3.0.1
beautifulsoup4==4.8.2
certifi==2024.12.14
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.1
compressed-rtf==1.0.6
decorator==5.1.1
Django==5.1.4
django-cors-headers==4.6.0
django-polymorphic==3.1.0
django-rest-polymorphic==0.1.10
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docx2txt==0.8
ebcdic==1.1.1
et_xmlfile==2.0.0
eval_type_backport==0.2.2
exceptiongroup==1.2.2
extract-msg==0.28.7
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
IMAPClient==2.1.0
joblib==1.4.2
jsonpath-python==1.0.6
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
lxml==5.3.0
mistralai==1.2.5
moviepy==2.1.1
msgpack==1.1.0
mypy-extensions==1.0.0
numba==0.60.0
numpy==2.0.2
olefile==0.47
openpyxl==3.1.5
packaging==24.2
pdfminer.six==20191110
pillow==10.4.0
platformdirs==4.3.6
pooch==1.8.2
proglog==0.1.10
psycopg2-binary==2.9.10
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.4
pydantic_core==2.27.2
PyJWT==2.10.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-pptx==0.6.23
pytz==2024.2
requests==2.32.3
scikit-learn==1.6.0
scipy==1.14.1
six==1.12.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.6
soxr==0.5.0.post1
SpeechRecognition==3.8.1
sqlparse==0.5.3
textract==1.6.5
threadpoolctl==3.5.0
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.3.0
xlrd==1.2.0
XlsxWriter==3.2.0
